+++ app/Http/Controllers/TagsController.php
@@ -104,6 +104,11 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count() > 0){
+            session()->flash('error','Tag cannot be deleted because it is associated to some post(s).');
+            return redirect()->back();
+        };
+
         $tag->delete();
 
         session()->flash('success','You have successfully deleted.');
